refactor(1013): extract DataCollectionManager to src/analytics/ (SC-001, position 25)
Extract DataCollectionManager (menu 76 continuous loop + menu 78 support
packages) from MistHelper.py to src/analytics/data_collection_manager.py.

- New: src/analytics/data_collection_manager.py (10 static methods, lazy mh
  import for ConfigUtils/OrgSiteExporter/OrgInventoryExporter/
  OrgDeviceStatsExporter/OrgAlarmEventExporter/GatewayTestExporter/
  CacheUtils/FilePathUtils).
- MistHelper.py: add re-export alias (Cat B position 25), replace 156-line
  class body with breadcrumb comment.
- pyproject.toml: extend coverage omit list.

Local gate: black + ruff + py_compile + --test (66/0/131 baseline held).
