+++ pyproject.toml
@@ -233,6 +233,7 @@ omit = [
     "setup.py",
     "*/site-packages/*",
     "src/maps/*",
+    "src/analytics/data_collection_manager.py",
     "src/api/api_data_fetcher.py",
     "src/export/const_definitions_exporter.py",
     "src/export/license_export_utils.py",
